simplifying the Container Manager
